fix: only cleanup files with field and collection data

+++ src/file-upload/data/repositories/file-uploads-repository-impl.ts
@@ -56,10 +56,12 @@ export class FileUploadsRepositoryImpl implements FileUploadsRepository {
 
     for await (const item of cursor) {
       const { collection, field, url, key, _id } = item;
-      const count = await db.collection(collection).findOne({ [field]: url });
-      if (!count) {
-        await this.fileUploadService.deleteFile(key);
-        await this.fileUploads.deleteOne({ _id });
+      if (collection && field) {
+        const count = await db.collection(collection).findOne({ [field]: url });
+        if (!count) {
+          await this.fileUploadService.deleteFile(key);
+          await this.fileUploads.deleteOne({ _id });
+        }
       }
     }
   }
